cuda_agent/requirements.txt
claude-agent-sdk>=0.1,<1.0
mcp>=1.20,<2.0
httpx>=0.27,<1.0
hydra-core>=1.3,<2.0
omegaconf>=2.3,<3.0
